1032 - Refactored aggregation util for  better efficiency.

+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -16,7 +16,7 @@ class EstimateIndCQCFilledPostsByJobRoleUtilsSchemas:
     )
     expected_aggregate_ascwds_worker_job_roles_per_establishment_schema = pl.Schema(
         list(aggregate_ascwds_worker_job_roles_per_establishment_schema.items())
-        + [(IndCQC.ascwds_job_role_counts, pl.Int64())]
+        + [(IndCQC.ascwds_job_role_counts, pl.UInt32())]
     )
 
     estimates_df_before_join_schema = pl.Schema(
